+++ src/Component/Http/Message.php
@@ -11,6 +11,8 @@
 		/** @var string[] */
 		const HTTP_METHODS = [ 'GET', 'HEAD', 'POST', 'PUT', 'DELETE', 'CONNECT', 'OPTIONS', 'TRACE', 'PATCH' ];
 
+		const HTTP_VERSIONS = [ '0.9', '1.0', '1.1', '2' ];
+
 		/** @var string[] */
 		const HTTP_STATUS = [
 			100 => 'Continue',