+++ src/Exception/ItemNotFoundException.php
@@ -1,0 +1,12 @@
+<?php
+
+	namespace Puzzle\Exception;
+
+	use \Psr\Container\NotFoundExceptionInterface;
+
+
+	class ItemNotFoundException extends \Exception implements NotFoundExceptionInterface {
+
+
+
+	}